requirements.txt
Flask
Flask-PyMongo
pymongo
Flask-RESTful
werkzeug
cerberus
flask-swagger-ui
PyJWT
